+++ lib/src/route_element.dart
@@ -69,12 +69,13 @@ abstract class VRouteElement {
   ///
   /// Also see:
   ///   * [VRouter.buildTransition] for default transitions for all routes
-  Widget Function(Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child)? get buildTransition;
+  Widget Function(
+          Animation<double> animation, Animation<double> secondaryAnimation, Widget child)?
+      get buildTransition;
 
   /// This is called before the url is updated but after all beforeLeave are called
   /// Note that it is only called if this [VRouteElement] is the last of the current route
-  ///
+  /// Use [newVRouteData] if you want information on the new route
   /// Return false if you don't want to redirect
   ///
   /// Note that you should consider the navigation cycle to
@@ -82,12 +83,14 @@ abstract class VRouteElement {
   ///
   /// Also see:
   ///   * [VRouter.beforeEnter] for global level beforeEnter
-  Future<bool> Function(BuildContext context, String? from, String to)?
+  Future<bool> Function(
+          BuildContext context, String? from, String to, VRouteData newVRouteData)?
       get beforeEnter;
 
   /// This is called before the url is updated if this [VRouteElement] is the
   /// last of the current route
-  ///
+  /// Use [newVRouteData] if you want information on the new route but be
+  /// careful, on the web newVRouteData is null when a user types a url manually
   /// Return false if you don't want to redirect
   ///
   /// [saveHistoryState] can be used to save a history state before leaving
@@ -101,7 +104,11 @@ abstract class VRouteElement {
   /// Also see:
   ///   * [VRouter.beforeLeave] for global level beforeLeave
   ///   * [VNavigationGuard.beforeLeave] for widget level beforeLeave
-  Future<bool> Function(BuildContext context, String? from, String to,
+  Future<bool> Function(
+      BuildContext context,
+      String? from,
+      String to,
+      VRouteData? newVRouteData,
       void Function(String state) saveHistoryState)? get beforeLeave;
 
   /// This is called after the url and the state is updated if this [VRouteElement]
@@ -222,23 +229,23 @@ class VStacked extends VRouteElement {
 
   /// See [VRouteElement.buildTransition]
   @override
-  final Widget Function(Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child)? buildTransition;
+  final Widget Function(
+          Animation<double> animation, Animation<double> secondaryAnimation, Widget child)?
+      buildTransition;
 
   /// See [VRouteElement.beforeEnter]
   @override
-  final Future<bool> Function(BuildContext context, String? from, String to)?
-      beforeEnter;
+  final Future<bool> Function(
+      BuildContext context, String? from, String to, VRouteData newVRouteData)? beforeEnter;
 
   /// See [VRouteElement.beforeLeave]
   @override
   final Future<bool> Function(BuildContext context, String? from, String to,
-      void Function(String state) saveHistoryState)? beforeLeave;
+      VRouteData? newVRouteData, void Function(String state) saveHistoryState)? beforeLeave;
 
   /// See [VRouteElement.afterEnter]
   @override
-  final void Function(BuildContext context, String? from, String to)?
-      afterEnter;
+  final void Function(BuildContext context, String? from, String to)? afterEnter;
 
   /// See [VRouteElement.onPop]
   @override
@@ -274,24 +281,20 @@ class VStacked extends VRouteElement {
         aliasesRegExp = (aliases != null)
             ? [
                 for (var alias in aliases)
-                  pathToRegExp(
-                      alias.startsWith('/') ? alias.substring(1) : alias)
+                  pathToRegExp(alias.startsWith('/') ? alias.substring(1) : alias)
               ]
             : null,
         parameters = <String>[],
-        aliasesParameters = (aliases != null)
-            ? List<List<String>>.filled(aliases.length, [])
-            : null,
+        aliasesParameters =
+            (aliases != null) ? List<List<String>>.filled(aliases.length, []) : null,
         navigatorKey = (subroutes != null &&
                 subroutes.isNotEmpty &&
-                subroutes.indexWhere((vChildClass) => vChildClass.isChild) !=
-                    -1)
+                subroutes.indexWhere((vChildClass) => vChildClass.isChild) != -1)
             ? GlobalKey<NavigatorState>()
             : null,
         heroController = (subroutes != null &&
                 subroutes.isNotEmpty &&
-                subroutes.indexWhere((vChildClass) => vChildClass.isChild) !=
-                    -1)
+                subroutes.indexWhere((vChildClass) => vChildClass.isChild) != -1)
             ? HeroController()
             : null,
         stateKey = GlobalKey<_RouteElementWidgetState>() {
@@ -393,23 +396,23 @@ class VChild extends VRouteElement {
 
   /// See [VRouteElement.buildTransition]
   @override
-  final Widget Function(Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child)? buildTransition;
+  final Widget Function(
+          Animation<double> animation, Animation<double> secondaryAnimation, Widget child)?
+      buildTransition;
 
   /// See [VRouteElement.beforeEnter]
   @override
-  final Future<bool> Function(BuildContext context, String? from, String to)?
-      beforeEnter;
+  final Future<bool> Function(
+      BuildContext context, String? from, String to, VRouteData newVRouteData)? beforeEnter;
 
   /// See [VRouteElement.beforeLeave]
   @override
   final Future<bool> Function(BuildContext context, String? from, String to,
-      void Function(String state) saveHistoryState)? beforeLeave;
+      VRouteData? newVRouteData, void Function(String state) saveHistoryState)? beforeLeave;
 
   /// See [VRouteElement.afterEnter]
   @override
-  final void Function(BuildContext context, String? from, String to)?
-      afterEnter;
+  final void Function(BuildContext context, String? from, String to)? afterEnter;
 
   /// See [VRouteElement.onPop]
   @override
@@ -445,24 +448,20 @@ class VChild extends VRouteElement {
         aliasesRegExp = (aliases != null)
             ? [
                 for (var alias in aliases)
-                  pathToRegExp(
-                      alias.startsWith('/') ? alias.substring(1) : alias)
+                  pathToRegExp(alias.startsWith('/') ? alias.substring(1) : alias)
               ]
             : null,
         parameters = <String>[],
-        aliasesParameters = (aliases != null)
-            ? List<List<String>>.filled(aliases.length, [])
-            : null,
+        aliasesParameters =
+            (aliases != null) ? List<List<String>>.filled(aliases.length, []) : null,
         navigatorKey = (subroutes != null &&
                 subroutes.isNotEmpty &&
-                subroutes.indexWhere((vChildClass) => vChildClass.isChild) !=
-                    -1)
+                subroutes.indexWhere((vChildClass) => vChildClass.isChild) != -1)
             ? GlobalKey<NavigatorState>()
             : null,
         heroController = (subroutes != null &&
                 subroutes.isNotEmpty &&
-                subroutes.indexWhere((vChildClass) => vChildClass.isChild) !=
-                    -1)
+                subroutes.indexWhere((vChildClass) => vChildClass.isChild) != -1)
             ? HeroController()
             : null,
         stateKey = GlobalKey<_RouteElementWidgetState>() {
@@ -532,8 +531,8 @@ class VChild extends VRouteElement {
 class VRouteRedirector extends VRouteElement {
   /// See [VRouteElement.beforeEnter]
   @override
-  final Future<bool> Function(BuildContext context, String? from, String to)?
-      beforeEnter;
+  final Future<bool> Function(
+      BuildContext context, String? from, String to, VRouteData newVRouteData)? beforeEnter;
 
   /// See [VRouteElement.path]
   @override
@@ -560,24 +559,23 @@ class VRouteRedirector extends VRouteElement {
   VRouteRedirector({
     required this.path,
     this.redirectTo,
-    Future<bool> Function(BuildContext context, String? from, String to)?
+    Future<bool> Function(
+            BuildContext context, String? from, String to, VRouteData newVRouteData)?
         beforeEnter,
     this.name,
     this.aliases,
   })  : assert(redirectTo != null || beforeEnter != null),
         assert(redirectTo == null || beforeEnter == null,
             'You should specify redirectTo OR beforeEnter but not both'),
-        pathRegExp =
-            pathToRegExp(path.startsWith('/') ? path.substring(1) : path),
+        pathRegExp = pathToRegExp(path.startsWith('/') ? path.substring(1) : path),
         aliasesRegExp = (aliases != null)
             ? [
                 for (var alias in aliases)
-                  pathToRegExp(
-                      alias.startsWith('/') ? alias.substring(1) : alias)
+                  pathToRegExp(alias.startsWith('/') ? alias.substring(1) : alias)
               ]
             : null,
         beforeEnter = beforeEnter ??
-            ((context, __, ___) async {
+            ((context, __, ___, ____) async {
               VRouterData.of(context).pushReplacement(redirectTo!);
               return false;
             });
@@ -600,13 +598,16 @@ class VRouteRedirector extends VRouteElement {
 
   /// Not implemented, this class is only for redirection
   @override
-  Future<bool> Function(BuildContext context, String? from, String to,
+  Future<bool> Function(
+      BuildContext context,
+      String? from,
+      String to,
+      VRouteData? newVRouteData,
       void Function(String state) saveHistoryState)? get beforeLeave => null;
 
   /// Not implemented, this class is only for redirection
   @override
-  void Function(BuildContext context, String? from, String to)?
-      get afterEnter => null;
+  void Function(BuildContext context, String? from, String to)? get afterEnter => null;
 
   /// Not implemented, this class is only for redirection
   @override
@@ -619,9 +620,8 @@ class VRouteRedirector extends VRouteElement {
   /// Not implemented, this class is only for redirection
   @override
   Widget Function(
-      Animation<double> animation,
-      Animation<double> secondaryAnimation,
-      Widget child)? get buildTransition => null;
+          Animation<double> animation, Animation<double> secondaryAnimation, Widget child)?
+      get buildTransition => null;
 
   /// Not implemented, this class is only for redirection
   @override